Coalesce span and move FeedbackBlock definition
